reset works & navigation changed
