+++ package.json
@@ -13,6 +13,7 @@
     "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
     "@react-native-async-storage/async-storage": "^1.19.1",
     "@react-native-picker/picker": "^2.4.10",
+    "@react-navigation/bottom-tabs": "^6.5.8",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
     "@react-navigation/stack": "^6.3.16",
@@ -21,6 +22,7 @@
     "lottie-react-native": "^6.0.1",
     "react": "18.2.0",
     "react-native": "^0.71.10",
+    "react-native-dotenv": "^3.4.9",
     "react-native-gesture-handler": "^2.12.0",
     "react-native-immersive": "^2.0.0",
     "react-native-rename": "^3.2.13",
