+++ src/Components/Productivity/AddTodos.js
@@ -10,15 +10,14 @@ import {
   TextInput,
   Dimensions,
 } from 'react-native';
-import React, {useState, useEffect} from 'react';
-import {bindActionCreators} from 'redux';
+import React, {useEffect} from 'react';
 import {
   updateAllTodos,
   updateTodoDescription,
   updateTodoTitle,
   updateTodoStatus,
   updateTodoPriority,
-} from '../Redux/Action-Creators';
+} from '../../Redux/Action-Creators';
 import {
   widthPercentageToDP as wp,
   heightPercentageToDP as hp,
@@ -26,88 +25,93 @@ import {
 import {connect} from 'react-redux';
 import Toast from 'react-native-toast-message';
 import {Picker} from '@react-native-picker/picker';
+import {bindActionCreators} from 'redux';
+import Snackbar from 'react-native-snackbar';
+import {ROOT_URL} from '../../Config/constants';
 
-const {width, height} = Dimensions.get('screen');
-const UpdateTodos = ({actions, navigation, route}) => {
-  const {jwtToken, id, todoTitle, todoDescription, todoStatus, todoPriority} =
-    route.params;
+
+const {width, height} = Dimensions.get('window');
+
+const AddTodos = ({
+  todoTitle,
+  todoDescription,
+  todoStatus,
+  todoPriority,
+  actions,
+  navigation,
+  route,
+}) => {
   useEffect(() => {
-    console.log('TODO STATUS -> ' + todoStatus);
-    console.log('TODO PRIORITY -> ' + todoPriority);
+    resetTodo();
   }, []);
-  const [title, setTitle] = useState(todoTitle);
-  const [description, setDescription] = useState(todoDescription);
-  const [priority, setPriority] = useState(todoPriority);
-  const [status, setStatus] = useState(todoStatus);
-  const handleTitle = newTitle => {
-    setTitle(newTitle);
-    actions.updateTodoTitle(newTitle);
-  };
-  const handleDescription = newDescription => {
-    setDescription(newDescription);
-    actions.updateTodoDescription(newDescription);
-  };
-  const handleStatus = newStatus => {
-    setStatus(newStatus);
-    actions.updateTodoStatus(newStatus);
-    console.log(newStatus);
-  };
-  const handlePriority = newPriority => {
-    setPriority(newPriority);
-    actions.updateTodoPriority(newPriority);
-    console.log(newPriority);
-  };
-
-  const handleUpdateTodo = async () => {
+  const {jwtToken} = route.params;
+  const handleAddTodo = async () => {
     try {
-      const res = await fetch(
-        `http://13.235.13.123:8082/user/updateTodos/${id}`,
-        {
-          method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${jwtToken}`,
-          },
-          body: JSON.stringify({
-            todoTitle: title,
-            todoDescription: description,
-            todoPriority: priority,
-            todoStatus: status,
-          }),
+      const res = await fetch(`${ROOT_URL}/user/api/v1/todos/add`, {
+        // 209 for linux , 154 pc
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${jwtToken}`,
         },
-      );
-      const updatedData = await res.json();
-      if (res.status === 200) {
+        body: JSON.stringify({
+          todoTitle: todoTitle,
+          todoDescription: todoDescription,
+          todoPriority: todoPriority,
+          todoStatus: todoStatus,
+        }),
+      });
+      const data = await res.json();
+      if (res.status === 201) {
         successToast();
         navigation.goBack();
-        console.log(updatedData);
+        console.log(data);
       } else if (res.status < 200 || res.status >= 300) {
-        console.log('ERROR ' + res.status + ' - ' + res.statusText);
+        console.log('ERROR ' + data.message);
+        Snackbar.show({
+          text: data.message,
+          duration: Snackbar.LENGTH_LONG,
+          backgroundColor: '#FFB800',
+          textColor: 'black',
+        });
       }
     } catch (error) {
       errorToast();
-      console.log('Error' + error);
+      console.log('Error ' + error);
+      Snackbar.show({
+        text: error,
+        duration: Snackbar.LENGTH_LONG,
+        backgroundColor: '#FFB800',
+        textColor: 'black',
+      });
     }
   };
 
   const successToast = () => {
     Toast.show({
       type: 'success',
-      text1: 'Successfully updated your Todo ',
+      text1: 'Successfully added your Todo ',
     });
   };
 
   const errorToast = () => {
     Toast.show({
       type: 'error',
-      text1: 'Error updating your Todo',
+      text1: 'Error adding your Todo',
     });
+  };
+
+  const resetTodo = () => {
+    actions.updateTodoTitle('');
+    actions.updateTodoDescription('');
+    actions.updateTodoStatus('');
+    actions.updateTodoPriority('');
   };
 
   return (
     <SafeAreaView style={styles.container}>
       <View style={styles.headerContainer}>
-        <Text style={styles.headerText}> Update To-Do </Text>
+        <Text style={styles.headerText}>Add To-Do </Text>
       </View>
       <ScrollView>
         <View>
@@ -116,9 +120,9 @@ const UpdateTodos = ({actions, navigation, route}) => {
               placeholder="Task Title"
               placeholderTextColor={'white'}
               style={styles.inputText}
-              value={title}
-              onChangeText={newTitle => {
-                handleTitle(newTitle);
+              value={todoTitle}
+              onChangeText={text => {
+                actions.updateTodoTitle(text);
               }}
             />
           </View>
@@ -129,18 +133,18 @@ const UpdateTodos = ({actions, navigation, route}) => {
                 placeholder="Task Brief"
                 placeholderTextColor={'grey'}
                 style={styles.inputText2}
-                value={description}
-                onChangeText={newDescription => {
-                  handleDescription(newDescription);
+                value={todoDescription}
+                onChangeText={text => {
+                  actions.updateTodoDescription(text);
                 }}
               />
             </View>
           </View>
           <View style={styles.pickerContainer}>
             <Picker
-              selectedValue={status}
-              onValueChange={newStatus => {
-                handleStatus(newStatus);
+              selectedValue={todoStatus}
+              onValueChange={item => {
+                actions.updateTodoStatus(item);
               }}>
               <Picker.Item label="Select Progress" value="" />
               <Picker.Item label="Completed" value="Completed" />
@@ -150,9 +154,9 @@ const UpdateTodos = ({actions, navigation, route}) => {
           </View>
           <View style={styles.pickerContainer2}>
             <Picker
-              selectedValue={priority}
-              onValueChange={newPriority => {
-                handlePriority(newPriority);
+              selectedValue={todoPriority}
+              onValueChange={item => {
+                actions.updateTodoPriority(item);
               }}>
               <Picker.Item label="Select Priority" value="" />
               <Picker.Item label="High" value="High" />
@@ -163,12 +167,8 @@ const UpdateTodos = ({actions, navigation, route}) => {
           </View>
         </View>
       </ScrollView>
-      <TouchableOpacity
-        style={styles.notesButton}
-        onPress={() => {
-          handleUpdateTodo();
-        }}>
-        <Text style={styles.notesText}>Update Todo</Text>
+      <TouchableOpacity style={styles.notesButton} onPress={handleAddTodo}>
+        <Text style={styles.notesText}>Save Todo</Text>
       </TouchableOpacity>
     </SafeAreaView>
   );
@@ -240,7 +240,6 @@ const styles = StyleSheet.create({
   },
 });
 
-//map state to props
 const mapStateToProps = state => ({
   todoTitle: state.addTodo.todoTitle,
   todoDescription: state.addTodo.todoDescription,
@@ -262,4 +261,4 @@ const mapDispatchToProps = dispatch => ({
   ),
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(UpdateTodos);
+export default connect(mapStateToProps, mapDispatchToProps)(AddTodos);
