+++ src/Components/Workspace/WorkSpaceHome.js
@@ -1,0 +1,11 @@
+import {View, Text} from 'react-native';
+
+const WorkSpaceHome = () => {
+  return (
+    <View>
+        <Text>Hello </Text>
+    </View>
+  );
+};
+
+export default WorkSpaceHome;
