+++ src/Config/colors.js
@@ -1,0 +1,21 @@
+export const DARKMODE = {
+    background:"#0a0700",
+    buttons:"#fcb40a",
+    buttonsText:"#ffffff",
+    headerText:"#ffb700",
+    labelText:"#c7c7c7",
+    placeholder:"#e8e6e6",
+    bottomNav:"#1f1f1f",
+    bottomNavText:"#FFFFFF",
+    bottomNavActive:"#313131"
+}
+
+export const LIGHTMODE = {
+    background:"#ffffff",
+    buttons:"#fcb40a",
+    buttonsText:"#000000",
+    headerText:"#0e0e0e",
+    labelText:"#c7c7c7",
+    placeholder:"#e8e6e6",
+    bottomNav:"#ececec"
+}
