+++ src/Config/constants.js
@@ -1,0 +1,1 @@
+export const ROOT_URL = 'http://192.168.1.20:8082';
